youBot joint setpoint update
The previous function getArmSetPointData and getArmSensorData were not getting synchronized data. To ensure this synchronization with the joint state data, they were implemented in the function already existing computeOODLSensorReadings.
Right now, there is an issue with the conversion from current to torque. (it has to do with the gearRatio and TorqueConstant values)...