Virtual trajectories estimation: filtering
The signal is filtered with a low pass filter (8 - 8.5Hz).
This filtering is not sufficient, a continuous element of the perturbations seems to remains.
A repetition of the isolated perturbation is filtered, then a linear interpolation is done between the begining of the perturbation, and a point 65ms later. This interpolation is substracted to the previous filtering to obtain the "continuous" part of the perturbation.
The continuous part is substracted to the original filtering (during the 65 ms of each perturbations), then the low pass filter is applied one last time to insure the continuity of the virtual trajectory.