Impedance identification live script
A matlab live script in under progress to explicit the methodologies proposed for impedance estimation.
The relative standard deviation criteria has been added to the impedance estimation, it allows to qualify the least square estimation on a parameter, and therfore can be interpreted as a confidence index.
And minor fixes.