Impedance live script
Last correction, the methodology brings the expected results when the virtual trajectory is known.
The ARX coefficient was solves by converting the discrete state space sys, to a discrete transfert function. The impedance parameters are then identified using the coefficient.
----
Diff trajectory minor fix (raise warning when the perturbation is too close to the end of the experiment and does not allow having a complete interval)
Impedance-data update, to allow separated statistic computation, next stage is to ad the ARX identification methodology that seems to bring better results